+++ src/App.js
@@ -1,19 +1,12 @@
 import React from "react";
-import "./App.css";
 import { Route, Switch } from "react-router-dom";
 
-import Authentification from "./components/Authentification/Authentification";
-import Toolbar from "./components/Toolbar/Toolbar";
+import Authentification from "./pages/Authentification";
 
-function App() {
-  return (
-    <div className="App" style={{ height: "100%" }}>
-      <Toolbar />
-      <Switch>
-        <Route exact path="/" component={Authentification} />
-      </Switch>
-    </div>
-  );
-}
+const App = () => (
+  <Switch>
+    <Route exact path="/" component={Authentification} />
+  </Switch>
+);
 
 export default App;
