+++ src/components/Layout/Toolbar/Toolbar.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import "../../styles/Toolbar.scss";
+import "./Toolbar.scss";
 
 const Toolbar = () => {
   return (
