+++ src/pages/Authentification/Authentification.scss
@@ -7,7 +7,7 @@
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  background: url("../img/mainBcg.jpg") center/cover no-repeat;
+  background: url("../../img/mainBcg.jpg") center/cover no-repeat;
 
   &:after {
     content: "";
